+++ canvas_sdk/v1/data/__init__.py
@@ -174,6 +174,7 @@ from .patient_consent import (
     PatientConsentRejectionCoding,
 )
 from .patient_group import PatientGroup, PatientGroupMember
+from .payment_card import PaymentCard
 from .payment_collection import PaymentCollection
 from .payor_specific_charge import PayorSpecificCharge
 from .plan import Plan
@@ -209,6 +210,7 @@ from .questionnaire import (
     InterviewQuestionnaireMap,
     InterviewQuestionResponse,
     Question,
+    QuestionEnablementCondition,
     Questionnaire,
     QuestionnaireQuestionMap,
     ResponseOption,
@@ -434,6 +436,7 @@ __all__ = __exports__ = (
     "PatientGroup",
     "PatientGroupMember",
     "PayorSpecificCharge",
+    "PaymentCard",
     "PaymentCollection",
     "Plan",
     "PluginCommand",
@@ -457,6 +460,7 @@ __all__ = __exports__ = (
     "ProtocolCurrent",
     "ProtocolOverride",
     "Question",
+    "QuestionEnablementCondition",
     "Questionnaire",
     "QuestionnaireQuestionMap",
     "ReasonForVisit",